separation of concerns / daily-hourly tempratures
